+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'blocs/bloc_navigator.dart';
 import 'blocs/bloc_theme.dart';
 import 'domains/blocs/bloc_session.dart';
 import 'domains/repositories/session_repository.dart';
@@ -9,21 +10,22 @@ import 'infrastructure/repositories/session_repository_impl.dart';
 import 'infrastructure/services/fake_service_session.dart';
 import 'shared/app_state_manager.dart';
 import 'shared/theme.dart';
-import 'views/splash_view.dart';
+import 'views/project_views_widget.dart';
 
 void main() {
   final BlocTheme blocTheme = BlocTheme();
   final ServiceSession serviceSession = FakeServiceSession();
 
-  // Debes proveer una implementación real de SessionRepository aquí
   final SessionRepository sessionRepository = SessionRepositoryImpl(
     SessionGatewayImpl(serviceSession),
   );
   final BlocSession blocSession = BlocSession(sessionRepository);
+  final BlocNavigator blocNavigator = BlocNavigator(blocSession);
   runApp(
     AppStateManager(
       blocTheme: blocTheme,
       blocSession: blocSession,
+      blocNavigator: blocNavigator,
       child: const MyApp(),
     ),
   );
@@ -39,7 +41,14 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: planningPokerTheme,
-      home: const SplashView(),
+      home: Builder(
+        builder: (BuildContext context) {
+          final BlocNavigator blocNavigator = AppStateManager.of(
+            context,
+          ).blocNavigator;
+          return ProjectViewsWidget(blocNavigator: blocNavigator);
+        },
+      ),
     );
   }
 }